@@ -1,20 +1,61 @@
-// This sample shows adding an action to an [AppBar] that opens a shopping cart.
-
 import 'package:flutter/material.dart';
-import 'package:nagano_gomi_oshirase2/calendar_page.dart';
-import 'package:nagano_gomi_oshirase2/notification_scheduler.dart';
-import 'package:nagano_gomi_oshirase2/preference.dart';
-import 'package:android_alarm_manager/android_alarm_manager.dart';
 
-void main() => runApp(MyApp());
+import 'calendar_page.dart';
+import 'notification_scheduler.dart';
+import 'preference.dart';
+import 'package:android_alarm_manager_plus/android_alarm_manager_plus.dart';
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await AndroidAlarmManager.initialize();
+  runApp(const MyApp());
+  NotificationScheduler.initialize();
+}
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
-  MyApp() {
-    AndroidAlarmManager.initialize();
-    NotificationScheduler.initialize();
-  }
+  const MyApp({super.key});
 
+  // This widget is the root of your application.
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        // This is the theme of your application.
+        //
+        // Try running your application with "flutter run". You'll see the
+        // application has a blue toolbar. Then, without quitting the app, try
+        // changing the primarySwatch below to Colors.green and then invoke
+        // "hot reload" (press "r" in the console where you ran "flutter run",
+        // or simply save your changes to "hot reload" in a Flutter IDE).
+        // Notice that the counter didn't reset back to zero; the application
+        // is not restarted.
+        primarySwatch: Colors.blue,
+      ),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+    );
+  }
+}
+
+class MyHomePage extends StatefulWidget {
+  const MyHomePage({super.key, required this.title});
+
+  // This widget is the home page of your application. It is stateful, meaning
+  // that it has a State object (defined below) that contains fields that affect
+  // how it looks.
+
+  // This class is the configuration for the state. It holds the values (in this
+  // case the title) provided by the parent (in this case the App widget) and
+  // used by the build method of the State. Fields in a Widget subclass are
+  // always marked "final".
+
+  final String title;
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -27,8 +68,9 @@ class MyApp extends StatelessWidget {
   }
 }
 
+
 class MyStatelessWidget extends StatelessWidget {
-  MyStatelessWidget({Key key}) : super(key: key);
+  MyStatelessWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -56,13 +98,13 @@ class MyStatelessWidget extends StatelessWidget {
 
 class AppBody extends StatelessWidget {
 
-  AppBody({Key key}) : super(key: key);
+  AppBody({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return new Scaffold(
       body: new Container(
-          child: new CalendarPage(),
+        child: new CalendarPage(),
       ),
     );
   }
