@@ -11,25 +11,24 @@ class CalendarItem {
 
 class CalendarUtil {
   static Future<Map<String, List<CalendarItem>>> getCalendar() async {
-    var response = await http.get("https://b-sw.co/nagano_gomi_calendar/gomi_calendar.json");
+    var response = await http.get(Uri.parse("https://b-sw.co/nagano_gomi_calendar/gomi_calendar.json"));
+    final o = <String, List<CalendarItem>>{};
     if (response.statusCode == 200) {
       String responseBody = convert.utf8.decode(response.bodyBytes);
       final jsonResponse = convert.jsonDecode(responseBody);
-      final o = new Map<String, List<CalendarItem>>();
       final dateFormatter = new DateFormat("yyyy-MM-dd");
       final today = dateFormatter.format(DateTime.now());
       for (var key in jsonResponse.keys) {
         final List<CalendarItem> h = jsonResponse[key]
-            .map((x) => new CalendarItem(x['date'], x['kind'], x['not_available']))
+            .map((x) => new CalendarItem(x['date'], x['kind'], x['not_available'] ?? false))
             .toList()
             .cast<CalendarItem>();
         h.sort((x, y) => x.date.compareTo(y.date));
         o[key] = h.where((x) => today.compareTo(x.date) <= 0).toList();
       }
-      return o;
     } else {
       print("Request failed with status: ${response.statusCode}.");
     }
-    return null;
+    return o;
   }
 }
